fix: write temporary cache file inside cache subdirectory
Prevents potential issues with permissions on the system default 
temporary directory returned by `sys_get_temp_dir()`.

+++ tests/Integration/IntegrationTest.php
@@ -41,7 +41,13 @@ abstract class IntegrationTest extends TestCase
             /** @var string $path */
             $path = $file->getRealPath();
 
-            unlink($path);
+            if ($file->isFile()) {
+                unlink($path);
+            }
+
+            if ($file->isDir()) {
+                rmdir($path);
+            }
         }
 
         rmdir($this->cacheDir);
